# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Inventory",
        "contact": {
            "email": "asutarko@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API My API"
        }
    ],
    "paths": {
...
        "/user/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login",
                "description": "Returns User data and token",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "abc@abc.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "login valid"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiMGRkN2YwZjQzMTJjYTUxYjYzNjc3MGE0ZDgyMzRiZTYyZWQ1OTU4ZjdhNTNiZTI3YWNiNzQ3ZDc5ZDkxNmRjMjBjM2M2MDY0ODllY2NlMzUiLCJpYXQiOjE2MjQ3MDQ2NDcsIm5iZiI6MTYyNDcwNDY0NywiZXhwIjoxNjU2MjQwNjQ3LCJzdWIiOiIxIiwic2NvcGVzIjpbXX0.OzFWNIiEFQ71X_2Xy-rH4RuBBOkYKEbs2toTtbSuzfF8Z01jguIGd5N-TPaMGATlIAC303hUcgvRnbuGb23CfJxH5UwaNJS1Q9hGQoMBEbr91xDf0xl_FduLWAgmOKD69xOl2_1kJ-z843o8Mov3zLMq190D0XB3EviJYtuo_Llo72hp11r8QAZoLL_4xGekfj-kOVSSR8rGFapSxvoDDQNccqba59kBtOikSDj_PvMQJ-YdRhAW5RAyqr8-xl5Qz5nnlbRO6UTeoSex1Fos3LiJxKhoJpHJXuNCN3NFKaoYok0qX9LnbdUlrS6HZTuo3bCbpTVyCJTXLdhbUGUSqAYgehotvDTaWD7KI62q-55k5jUM9BH45mtZStVSabDsy2F8eIsF1AepQTVN-2NR0056bYcVdw1IuF7xdySpmJm45ptcNGel3MnqBJpKlbpqF8oQ9gYYaa3tu7d3VeN-AelQgj9bUzAArAp4bone8r7x6s_Q0NwXKnmV2HMO_3cZ_shfdRfxy1L3BSinXjaVX3AlYgDM67DYf8HrHy5sBvMjiTB-6AIrbhAjoZP0sM3c8WZipXmG87KilniDRP3wQ5CRmeD918PARwhyd24kMeLjipomvDuv6r_86N4Dwm0UMHLmOHkNxpsGYUnAgkJRsh3TTG4dwZ8eflwfugt0Pfc"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocess",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "email kosong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Registration",
                "description": "Returns User data",
                "operationId": "Register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ani"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jalan Pramuka no 100"
                                    },
                                    "email": {
                                        "type": "email",
                                        "example": "abc@abc.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Register Berhasil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocess",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "nama Kosong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}